requirements.txt
websockets>=12.0
pydantic>=2.5.0
asyncio-mqtt>=0.16.0
aiofiles>=24.1.0
Pillow>=10.0.0
requests>=2.31.0
lxml>=4.9.0
psutil>=5.9.0
python-dotenv>=1.0.0
click>=8.1.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
xmltodict>=0.13.0
beautifulsoup4>=4.12.0
